fix: resolve merge conflicts and integrate ticket admin dashboard to the existing admin dashbord
- Fix security config duplicate method definition
- Resolve frontend TypeScript errors and prop mismatches
- Update backend API endpoint configuration
- Add ticket management to admin dashboard

+++ backend/src/main/java/com/project/smart_campus_operationhub/services/TicketAttachmentService.java
@@ -53,7 +53,7 @@ public class TicketAttachmentService {
         attachment.setTicket(ticket);
 
         Users user = new Users();
-        user.setId(Long.valueOf(userId));
+        user.setId(userId);
         attachment.setUploadedBy(user);
 
         attachment.setFileName(file.getOriginalFilename());
@@ -85,7 +85,7 @@ public class TicketAttachmentService {
         dto.setFilePath(att.getFilePath());
         dto.setFileType(att.getFileType());
         dto.setUploadedById(
-                att.getUploadedBy() != null ? att.getUploadedBy().getId().intValue() : null);
+                att.getUploadedBy() != null ? att.getUploadedBy().getId() : null);
         dto.setUploadedAt(att.getUploadedAt());
 
         return dto;
